gameplay small change, scenes files added

+++ Source/Components/DrawComponents/DrawSpriteWColorEffect.cpp
@@ -7,7 +7,7 @@
 
 
 DrawSpriteWColorEffect::DrawSpriteWColorEffect(Actor *owner, const std::string &texturePath, int width, int height,
-                                               int drawOrder, int r_, int g_, int b_, int a_) //chamando o construtor do DrawSprite, e desativando o efeito inicialmente
+                                               int drawOrder, int r_, int g_, int b_, float a_) //chamando o construtor do DrawSprite, e desativando o efeito inicialmente
                                                : DrawSpriteComponent(owner, texturePath, width, height, drawOrder), active(false),
                                                r(r_),
                                                g(g_),
@@ -31,7 +31,7 @@ void DrawSpriteWColorEffect::Draw(SDL_Renderer *renderer) {
     }
 }
 
-void DrawSpriteWColorEffect::SetColorEffect(int r_, int g_, int b_, int a_) {
+void DrawSpriteWColorEffect::SetColorEffect(int r_, int g_, int b_, float a_) {
     r = r_;
     g = g_;
     b = b_;
